factor out yaml processor as "camel" module
